Conflicts:
	Migrations/Cats.DatabaseSchema/Cats.DatabaseSchema.sqlproj
	Migrations/Cats.DatabaseSchema/EarlyWarning/Tables/ReliefRequisition.sql
	Migrations/Cats.DatabaseSchema/Procurement/Tables/BidDetail.sql
	Migrations/Cats.DatabaseSchema/Procurement/Tables/TransportBidPlanDetail.sql
	Migrations/Cats.DatabaseSchema/dbo/Tables/Bid.sql
	Migrations/Cats.DatabaseSchema/dbo/Tables/UserHub.sql
	Models/Cats.Models/TransportBidPlan.cs
	Web/Areas/Procurement/Controllers/RFQController.cs
	Web/Areas/Procurement/Models/RfqViewModel.cs
	Web/Areas/Procurement/Views/RFQ/Details.cshtml
	Web/Cats.csproj

+++ Models/Cats.Models/BidDetail.cs
@@ -11,8 +11,8 @@ namespace Cats.Models
         public int RegionID { get; set; }
         public decimal AmountForReliefProgram { get; set; }
         public decimal AmountForPSNPProgram { get; set; }
-        public float BidDocumentPrice { get; set; }
-        public float CPO { get; set; }
+        public decimal BidDocumentPrice { get; set; }
+        public decimal CPO { get; set; }
 
         #region Navigation Properties
